small refactor of file structure
