+++ .github/workflows/preview.yml
@@ -131,10 +131,7 @@ jobs:
         run: pnpm exec playwright install --with-deps
 
       - name: Run Playwright tests
-        run: pnpm exec playwright test playwright/global.spec.ts --workers 6
-
-      - name: Run Playwright tests
-        run: pnpm exec playwright test playwright/versioned-docs.spec.ts
+        run: pnpm exec playwright test playwright/smoke --workers 6
 
       - uses: actions/upload-artifact@v4
         if: always()
