+++ playwright/smoke/favicon.spec.ts
@@ -1,13 +1,5 @@
-import { expect, test } from "@playwright/test";
-import { getPlaywrightTestUrls } from "./test-runner";
-
-const existenceUrls = getPlaywrightTestUrls("existence");
-existenceUrls.forEach((testUrl) => {
-    test(`Check if ${testUrl} is online`, async ({ page }) => {
-        const response = await page.goto(testUrl);
-        expect(response?.status()).toBe(200);
-    });
-});
+import test, { expect } from "@playwright/test";
+import { getPlaywrightTestUrls } from "../test-runner";
 
 const faviconUrls = getPlaywrightTestUrls("favicon");
 faviconUrls.forEach((testUrl) => {
